+++ lib/assent/strategies/gitlab.ex
@@ -1,6 +1,6 @@
 defmodule Assent.Strategy.Gitlab do
   @moduledoc """
-  Gitlab OAuth 2.0 strategy.
+  Gitlab OpenID Connect strategy.
 
   ## Usage
 
@@ -12,30 +12,14 @@ defmodule Assent.Strategy.Gitlab do
 
   See `Assent.Strategy.OAuth2` for more.
   """
-  use Assent.Strategy.OAuth2.Base
+  use Assent.Strategy.OIDC.Base
 
   @impl true
   def default_config(_config) do
     [
       base_url: "https://gitlab.com",
-      authorize_url: "/oauth/authorize",
-      token_url: "/oauth/token",
-      user_url: "/api/v4/user",
-      authorization_params: [scope: "api read_user read_registry"],
-      auth_method: :client_secret_post
+      authorization_params: [scope: "email profile"],
+      client_authentication_method: "client_secret_post"
     ]
-  end
-
-  @impl true
-  def normalize(_config, user) do
-    {:ok,
-     %{
-       "sub" => user["id"],
-       "name" => user["name"],
-       "preferred_username" => user["username"],
-       "picture" => user["avatar_url"],
-       "email" => user["email"],
-       "email_verified" => not is_nil(user["confirmed_at"])
-     }}
   end
 end